// dashboard/src/types/apps/File/transactionDefinitionTypes.ts
export type TransactionDefinitionType = {
    id: string;
    transactionCode: string;
    transactionName: string;
    shortName: string;
    transactionType: string;
    updateType: string;
    permanent: string;
    taxable: string;
    unTaxableLimit: string;
    affectByLeave: string;
    leaveDays: string;
    affectBackPayroll: string;
    affectBeneficiary: string;
    transactionGroup: string;
    glEntryBy: string;
    directAccount: string;
    contractGLAccount: string;
    transactionTypeName: string;
    updateTypeName: string;
    transactionGroupName: string;
  }